requirements.txt
certifi==2020.12.5
click==8.0.0
colorama==0.4.4
dataclasses==0.6
dnspython==2.1.0
Flask==2.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pymongo==3.11.4
python-dotenv==0.17.1
Werkzeug==2.0.0
wincertstore==0.2
gunicorn==20.1.0